+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Next minor version
+
+- Add `proxy_protocol_use_cn_as_username` feature which for `proxy_protocol`
+  enabled listeners enable or disable using the common name forwarded by the
+  PROXY protocol instead of the MQTT username. To ensure backward compatibility
+  this setting is enabled by default. This feature was kindly contributed by
+  SoftAtHome (https://softathome.com/).
+
 ## Not yet released
 
 - New `vmq-admin` command to forcefully disconnect and cleanup sessions.